+++ client/src/main.jsx
@@ -2,6 +2,7 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
 import { AuthProvider } from './context/AuthContext.jsx'
+import { EditorProvider } from './context/EditorContext.jsx'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
 const queryClient = new QueryClient()
@@ -9,7 +10,9 @@ const queryClient = new QueryClient()
 createRoot(document.getElementById('root')).render(
   <QueryClientProvider client={queryClient}>
     <AuthProvider>
-      <App />
+      <EditorProvider>
+        <App />
+      </EditorProvider>
     </AuthProvider>
   </QueryClientProvider>
 )